Added whitelist command, fixed bug with censor and discord formatting
The wordfilter.py cog is STILL NON-FUNCTIONAL, but it's been left in just in case someone wants to take a crack at it. For now, the whitelist command is in bot.py.

wordfilter.py has been added to the files that COVERAGE WILL IGNORE.

The default whitelist (default_whitelist.txt) IS NOT CURRENTLY BEING LOADED. It's just there for future testing/implementations.

Added methods to profanity_helper.py to get around some weird bugs with the whitelist not loading properly and discord formatting everything.